Change BigQuerySinkTask
